+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     * render\_table\_content
     * render\_table\_js
 
+* Ignore HTML when filtering href columns
+
 ## v0.1.0
 
 First commit for codebase.