+++ CHANGELOG.md
@@ -15,13 +15,15 @@
 
 ## v0.1.1
 
-* New column: DatetimeColumn
+* New column:
+
+    * DatetimeColumn
+
 * New options to disable add-ons
 
     * disable_search
     * disable_info
     * disable_pagination
-    * disable_length_menu
 
 * Fix bug about filtering href columns, ignore text of HTML tag.